[homeconnect] Moved temperature command handling to abstract class and use channel type 'system.power' instead of custom type
